requirements.txt
boto3==1.20.54
python-dotenv==0.19.2
PyYAML==6.0